Fix double to float type cast error on Mac

+++ Source/DeviceThread.cpp
@@ -692,7 +692,7 @@ void DeviceThread::updateSettings(OwnedArray<ContinuousChannel>* continuousChann
         "description",
         "identifier",
 
-        evalBoard->getSampleRate()
+        static_cast<float>(evalBoard->getSampleRate())
 
     };
